style: upgrade header styles and homepage styles
