+++ .github/workflows/testnet-package.yaml
@@ -1,7 +1,7 @@
 name: Build and Push Image
 on:
   push:
-    branches: ['dev', 'feature/*', 'fix/*']
+    branches: ['dev', 'feature/*', 'fix/*', 'style/*']
 jobs:
   build-and-push-image:
     runs-on: eforest-runner
