+++ src/components/Header/style.module.css
@@ -13,15 +13,11 @@
 }
 
 .marketplace-header .nav-text {
-  @apply text-[var(--color-header)] not-italic font-semibold text-[18px] leading-[27px] cursor-pointer;
-}
-
-.marketplace-header .nav-text:hover {
-  @apply text-[var(--fill-white-1)];
+  @apply text-textPrimary hover:!text-textSecondary not-italic font-semibold text-[18px] leading-[27px] cursor-pointer;
 }
 
 .marketplace-header .nav-text.text-select {
-  @apply text-[var(--brand-light)];
+  @apply text-[var(--line-dash)];
 }
 
 .marketplace-header .icon-btn-wrap {
@@ -37,6 +33,7 @@
 }
 .marketplace-header.mobile-header-wrapper .forest-logo svg {
   @apply scale-[0.58];
+  transform-origin: 0 50%;
 }
 
 .marketplace-header .icon-btn-wrap :global .ant-space-item {
@@ -72,12 +69,35 @@
   @apply cursor-pointer;
 }
 
+.marketplace-header :global .ant-dropdown {
+  box-shadow: 0px 4px 20px 0px rgba(64, 64, 64, 0.1);
+}
+
 .mobile-header-wrapper {
   @apply py-[10px] px-[16px];
 }
 
 .marketplace-header.mobile-header-wrapper {
   @apply h-[62.4px];
+}
+
+.marketplace-header .menu {
+  :global .ant-dropdown-menu {
+    @apply !p-[8px];
+  }
+
+  :global .ant-dropdown-menu-item {
+    @apply w-[188px] !h-[52px] !px-[0px]  !border-0;
+
+    .ant-dropdown-menu-title-content {
+      @apply font-medium text-[16px] !px-[8px] w-full h-full hover:bg-fillHoverBg rounded-[8px];
+
+      span,
+      a {
+        @apply hover:!text-textPrimary;
+      }
+    }
+  }
 }
 
 .header-wallet-custom {
@@ -109,11 +129,11 @@
 
 .white-header {
   .marketplace-header {
-    @apply bg-white;
+    @apply bg-white border border-solid border-lineBorder;
   }
 
   .nav-text {
-    @apply text-textPrimary;
+    @apply text-textPrimary hover:!text-textSecondary;
   }
 
   span {
@@ -138,7 +158,7 @@
   }
 
   .nav-text {
-    @apply text-white;
+    @apply text-white hover:!text-[var(--line-dash)];
   }
 
   span {
