+++ src/pagesComponents/Home/components/Trending/styles.module.css
@@ -5,7 +5,8 @@
 }
 
 .CollectionItem {
-  @apply inline-block pr-[20px] w-[240px] mdb:w-[25%] mc:w-[20%];
+  /* @apply inline-block w-[240px] mdb:w-[25%] mc:w-[20%]; */
+  @apply inline-block w-full;
 
   transition: all 0.25s ease-in;
 
@@ -33,4 +34,12 @@
   @apply absolute top-[50%] translate-y-[-50%] right-[-20px] z-20;
 
   backdrop-filter: blur(5px);
+}
+
+.CollectionItem_wrapper {
+  @apply relative w-full rounded-lg overflow-hidden border-[1px] border-solid border-[var(--line-border)];
+}
+
+.CollectionItem_wrapper:hover {
+  box-shadow: 0px 4px 20px 0px rgba(64, 64, 64, 0.1);
 }
