+++ src/pagesComponents/Home/styles.module.css
@@ -1,3 +1,3 @@
 .wrapper-width {
-  @apply mdl:w-[calc(100vw-80px)] w-[calc(100vw-48px)];
+  @apply w-full;
 }
